+++ src/__tests__/__fakes__/FakeComponentWithArgs.fake.ts
@@ -1,12 +1,18 @@
 export default class FakeComponentWithArgs {
-  private _args: any
+  private _arg1: number;
+  private _arg2: string;
 
-  constructor(args: number) {
-    this._args = args
+  constructor(arg1: number, arg2: string) {
+    this._arg1 = arg1;
+    this._arg2 = arg2;
   }
 
-  public get args() {
-    return this._args
+  public get arg1() {
+    return this._arg1;
+  }
+
+  public get arg2() {
+    return this._arg2;
   }
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function