+++ src/__tests__/__fakes__/FakeSystem.ts
@@ -1,30 +1,24 @@
-import {IEntity} from '../../entity/types'
-import {System} from '../../system'
-import FakeComponentWithArgs from './FakeComponentWithArgs.fake'
-import FakeComponentWithNoArgs from './FakeComponentWithNoArgs.fake'
+import { IEntity } from "../../entity/types";
+import { System } from "../../system";
+import FakeComponentWithArgs from "./FakeComponentWithArgs.fake";
+import FakeComponentWithNoArgs from "./FakeComponentWithNoArgs.fake";
 
 /* eslint-disable @typescript-eslint/no-empty-function */
 
-export type TFakeSystemWithNoArgsComponents = [
-  FakeComponentWithArgs,
-  FakeComponentWithNoArgs
-]
+export type TFakeSystemComponents = [
+  typeof FakeComponentWithArgs,
+  typeof FakeComponentWithNoArgs
+];
 
-export default class FakeSystem extends System<
-  TFakeSystemWithNoArgsComponents,
-  any
-  > {
+export default class FakeSystem extends System<TFakeSystemComponents, never> {
   constructor() {
-    super()
-    this.setComponents(FakeComponentWithArgs)
+    super();
+    this.setComponents(FakeComponentWithArgs, FakeComponentWithNoArgs);
   }
 
-  public update(
-    _entity: IEntity<TFakeSystemWithNoArgsComponents>,
-    _dependencies: any
-  ): void { }
+  public update(_entity: IEntity<TFakeSystemComponents>): void {}
 
-  public async init() { }
+  public async init() {}
 
-  public preUpdate() { }
+  public preUpdate() {}
 }