+++ src/system/index.ts
@@ -1,6 +1,6 @@
-import {TComponentBase} from '../types'
-import System from './System'
+import { TAnyConstructor } from "../types";
+import System from "./System";
 
-export {System}
+export { System };
 
-export type ISystem<T extends TComponentBase[]> = import('./types').ISystem<T>
+export type ISystem<T extends TAnyConstructor[]> = import("./types").ISystem<T>;