+++ src/utils.ts
@@ -1,6 +1,6 @@
-import {IWorld} from './world/types'
-import World from './world/World'
+import { IWorld } from "./world/types";
+import World from "./world/World";
 
-export const createWorld = <T extends object>(
+export const createWorld = <T extends Record<string, unknown>>(
   dependencies: T
-): IWorld<T> => new World(dependencies)
+): IWorld<T> => new World(dependencies);